requirements.txt
discord
discord.py[voice]
requests
markdown
dnspython
aiohttp
aiofiles
asyncio
better-profanity
psutil
git+https://github.com/Rapptz/discord-ext-menus
python-dotenv
apscheduler
pyvips
youtube-dl
humanize
PyNaCl
wheel
ffmpeg
cffi
pillow